move fixed-point logic code into fixed_point module
